+++ NEWS.md
@@ -1,11 +1,17 @@
-# Version 0.2.0 (In development)
+# Version 0.2.5 (Current GitHub Version)
+* Fixed bug in `dExponentialGrowth()` causing memory allocation problems.
+* Added alternative parametrisation for logistic growth model (`dLogisticGrowth2()`)
+* Updated plot functions to allow user defined x and y labels.
+* Added option to specify line width in `modelPlot()`  
+
+# Version 0.2.1 (CRAN Version)
 * NEW Distributions:
   * Trapezoidal Distribution.
   * Asymmetric Laplace Distribution
 * Updated version of `modelPlot()` function.
 * Fixed error in `compare.models()` following changes in WAIC storing after nimble version 0.12.x
 
-# Version 0.1.2 (CRAN Version)
+# Version 0.1.2
 * Fixes an installation error affecting Oracle Solaris 10, x86, 32 bit, R release, Oracle Developer Studio 12.6 
 
 # Version 0.1.1